+++ clang-tools-extra/include-cleaner/unittests/TypesTest.cpp
@@ -34,12 +34,13 @@ TEST(RecordedIncludesTest, Match) {
   FileEntryRef B = FM.getVirtualFileRef("/path/b", /*Size=*/0, time_t{});
 
   Includes Inc;
-  Inc.add(Include{"a", A, SourceLocation(), 1});
-  Inc.add(Include{"a2", A, SourceLocation(), 2});
-  Inc.add(Include{"b", B, SourceLocation(), 3});
-  Inc.add(Include{"vector", B, SourceLocation(), 4});
-  Inc.add(Include{"vector", B, SourceLocation(), 5});
-  Inc.add(Include{"missing", std::nullopt, SourceLocation(), 6});
+  Inc.add(Include{"a", A, SourceLocation(), SourceLocation(), 1});
+  Inc.add(Include{"a2", A, SourceLocation(), SourceLocation(), 2});
+  Inc.add(Include{"b", B, SourceLocation(), SourceLocation(), 3});
+  Inc.add(Include{"vector", B, SourceLocation(), SourceLocation(), 4});
+  Inc.add(Include{"vector", B, SourceLocation(), SourceLocation(), 5});
+  Inc.add(
+      Include{"missing", std::nullopt, SourceLocation(), SourceLocation(), 6});
 
   EXPECT_THAT(Inc.match(A), ElementsAre(line(1), line(2)));
   EXPECT_THAT(Inc.match(B), ElementsAre(line(3), line(4), line(5)));
@@ -55,8 +56,9 @@ TEST(RecordedIncludesTest, MatchVerbatim) {
   // By default, a verbatim header only matches includes with the same spelling.
   auto Foo =
       FM.getVirtualFileRef("repo/lib/include/rel/foo.h", /*Size=*/0, time_t{});
-  Inc.add(Include{"lib/include/rel/foo.h", Foo, SourceLocation(), 1});
-  Inc.add(Include{"rel/foo.h", Foo, SourceLocation(), 2});
+  Inc.add(Include{"lib/include/rel/foo.h", Foo, SourceLocation(),
+                  SourceLocation(), 1});
+  Inc.add(Include{"rel/foo.h", Foo, SourceLocation(), SourceLocation(), 2});
   EXPECT_THAT(Inc.match(Header("<rel/foo.h>")), ElementsAre(line(2)));
 
   // A verbatim header can match another spelling if the search path
@@ -65,15 +67,17 @@ TEST(RecordedIncludesTest, MatchVerbatim) {
       FM.getVirtualFileRef("repo/lib/include/rel/bar.h", /*Size=*/0, time_t{});
   Inc.addSearchDirectory("repo/");
   Inc.addSearchDirectory("repo/lib/include");
-  Inc.add(Include{"lib/include/rel/bar.h", Bar, SourceLocation(), 3});
-  Inc.add(Include{"rel/bar.h", Bar, SourceLocation(), 4});
+  Inc.add(Include{"lib/include/rel/bar.h", Bar, SourceLocation(),
+                  SourceLocation(), 3});
+  Inc.add(Include{"rel/bar.h", Bar, SourceLocation(), SourceLocation(), 4});
   EXPECT_THAT(Inc.match(Header("<rel/bar.h>")),
               UnorderedElementsAre(line(3), line(4)));
 
   // We don't apply this logic to system headers, though.
   auto Vector =
       FM.getVirtualFileRef("repo/lib/include/vector", /*Size=*/0, time_t{});
-  Inc.add(Include{"lib/include/vector", Vector, SourceLocation(), 5});
+  Inc.add(Include{"lib/include/vector", Vector, SourceLocation(),
+                  SourceLocation(), 5});
   EXPECT_THAT(Inc.match(Header(*tooling::stdlib::Header::named("<vector>"))),
               IsEmpty());
 }
@@ -88,14 +92,14 @@ TEST(RecordedIncludesTest, MatchVerbatimMixedAbsoluteRelative) {
       FM.getVirtualFileRef("/working/rel1/rel2/foo.h", /*Size=*/0, time_t{});
   Inc.addSearchDirectory("rel1");
   Inc.addSearchDirectory("rel1/rel2");
-  Inc.add(Include{"rel2/foo.h", Foo, SourceLocation(), 1});
+  Inc.add(Include{"rel2/foo.h", Foo, SourceLocation(), SourceLocation(), 1});
   EXPECT_THAT(Inc.match(Header("<foo.h>")), IsEmpty());
 
   Inc = Includes{};
   auto Bar = FM.getVirtualFileRef("rel1/rel2/bar.h", /*Size=*/0, time_t{});
   Inc.addSearchDirectory("/working/rel1");
   Inc.addSearchDirectory("/working/rel1/rel2");
-  Inc.add(Include{"rel2/bar.h", Bar, SourceLocation(), 1});
+  Inc.add(Include{"rel2/bar.h", Bar, SourceLocation(), SourceLocation(), 1});
   EXPECT_THAT(Inc.match(Header("<bar.h>")), IsEmpty());
 }